Agrego ultimas funciones de la rutina principal

+++ src/main/java/Estado.java
@@ -1,5 +1,5 @@
 public class Estado {
 
-    public int beneficio = 0;
+    public double beneficio = 0;
     public int suscriptores = 70000;
 }
